:bug: :fire: gatsby-node cannot be use as typescript
