+++ gatsby-ssr.js
@@ -1,9 +1,5 @@
-require('source-map-support').install()
-require('ts-node').register({
-  compilerOptions: {
-    module: 'commonjs',
-    target: 'es2017',
-  },
-})
+import App from './src/app/components'
 
-exports.wrapPageElement = require('./src/server/wrapPageElement').wrapPageElement
+export const wrapPageElement = ({ element, props }) => {
+  return <App {...props}>{element}</App>
+}
