+++ src/core/components/reader.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ReaderComponent: React.FC = props => {
+  return (
+    <>OK</>
+  )
+}
+
+export default ReaderComponent
