--- src/server/@types/IDatabaseCode.ts
@@ -1,4 +1,0 @@
-export type IDatabaseCode = number | {
-  code: number
-  exclude?: number[]
-}
