--- src/server/@types/IDatabaseTag.ts
@@ -1,7 +1,0 @@
-import { ITagType } from '../../core/@types/ITagType'
-
-export interface IDatabaseTag {
-  color: string
-  name: ITagType
-  prefix: string
-}
