--- src/server/onClientEntry.ts
@@ -1,5 +1,0 @@
-export const onClientEntry = () => {
-  if (process.env.NODE_ENV !== `production`) {
-    require(`preact/debug`)
-  }
-}
