--- src/server/onCreatePage.ts
@@ -1,8 +1,0 @@
-export const onCreatePage = async ({page, actions}: any) => {
-  const {createPage} = actions
-
-  if (page.path.match(/^\/g/)) {
-    page.matchPath = '/g/*'
-    createPage(page)
-  }
-}
