--- src/server/onCreateWebpackConfig.ts
@@ -1,15 +1,0 @@
-import { CreateWebpackConfigArgs } from 'gatsby'
-
-export const onCreateWebpackConfig = ({ actions, stage }: CreateWebpackConfigArgs) => {
-  if (!stage.startsWith('develop')) {
-    actions.setWebpackConfig({
-      resolve: {
-        alias: {
-          react: `preact/compat`,
-          "react-dom": `preact/compat`,
-          "react-dom/server": `preact/compat`,
-        },
-      },
-    })
-  }
-}
