+++ src/templates/hentai/listing/components/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const HentaiListingComponent: React.FC = props => {
+  return (
+    <>OK</>
+  )
+}
+
+export default HentaiListingComponent
