+++ src/templates/hentai/viewing/components/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const HentaiViewingComponent: React.FC = props => {
+  return (
+    <>OK</>
+  )
+}
+
+export default HentaiViewingComponent
