+++ src/templates/tag/listing/components/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const TagListingComponent: React.FC = props => {
+  return (
+    <>OK</>
+  )
+}
+
+export default TagListingComponent
