+++ src/templates/tag/viewing/components/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const TagViewingComponent: React.FC = props => {
+  return (
+    <>OK</>
+  )
+}
+
+export default TagViewingComponent
